상세 기능 - Detail Page 컴포넌트 분리
